+++ src/Products/PloneMeeting/MeetingItem.py
@@ -110,6 +110,7 @@ from Products.PloneMeeting.config import WriteMarginalNotes
 from Products.PloneMeeting.content.meeting import Meeting
 from Products.PloneMeeting.events import item_added_or_initialized
 from Products.PloneMeeting.ftw_labels.utils import compute_labels_access
+from Products.PloneMeeting.ftw_labels.utils import get_labels
 from Products.PloneMeeting.interfaces import IMeetingItem
 from Products.PloneMeeting.interfaces import IMeetingItemWorkflowActions
 from Products.PloneMeeting.interfaces import IMeetingItemWorkflowConditions
@@ -138,6 +139,7 @@ from Products.PloneMeeting.utils import getFieldVersion
 from Products.PloneMeeting.utils import getWorkflowAdapter
 from Products.PloneMeeting.utils import hasHistory
 from Products.PloneMeeting.utils import is_editing
+from Products.PloneMeeting.utils import is_proposing_group_editor
 from Products.PloneMeeting.utils import isPowerObserverForCfg
 from Products.PloneMeeting.utils import ItemDuplicatedEvent
 from Products.PloneMeeting.utils import ItemDuplicatedToOtherMCEvent
@@ -476,8 +478,8 @@ class MeetingItemWorkflowConditions(object):
            self.context.hasMeeting():
             meeting = self.context.getMeeting()
             if meeting.date < datetime.now():
-                if not self.context.fieldIsEmpty('decision') or not \
-                   self.context.fieldIsEmpty('motivation'):
+                if not fieldIsEmpty('decision', self.context) or \
+                   not fieldIsEmpty('motivation', self.context):
                     res = True
                 else:
                     itemNumber = self.context.getItemNumber(relativeTo='meeting',
@@ -1576,7 +1578,7 @@ schema = Schema((
         default_output_type="text/x-html-safe",
         optional=True,
         # we use WriteMarginalNotes so MeetingManagers may edit votesResult
-        # when item is decided but as field in not in
+        # when item is decided but as field is not in
         # MeetingItem._bypass_meeting_closed_check_for it will not be quick editable
         # when the meeting is closed
         write_permission=WriteMarginalNotes,
@@ -1717,6 +1719,40 @@ schema = Schema((
         optional=True,
         read_permission=WriteInternalNotes,
         write_permission=WriteInternalNotes,
+    ),
+    TextField(
+        name="neededFollowUp",
+        allowable_content_types=("text/html",),
+        widget=RichWidget(
+            condition="python: here.adapted().show_field('neededFollowUp')",
+            label="Neededfollowup",
+            label_msgid="PloneMeeting_label_neededFollowUp",
+            description="NeededFollowUp",
+            description_msgid="needed_follow_up_descr",
+            i18n_domain="PloneMeeting",
+        ),
+        default_content_type="text/html",
+        default_output_type="text/x-html-safe",
+        searchable=True,
+        optional=True,
+        write_permission=View,
+    ),
+    TextField(
+        name="providedFollowUp",
+        allowable_content_types=("text/html",),
+        widget=RichWidget(
+            condition="python: here.adapted().show_field('providedFollowUp')",
+            label="Providedfollowup",
+            label_msgid="PloneMeeting_label_providedFollowUp",
+            description="ProvidedFollowUp",
+            description_msgid="provided_follow_up_descr",
+            i18n_domain="PloneMeeting",
+        ),
+        default_content_type="text/html",
+        default_output_type="text/x-html-safe",
+        searchable=True,
+        optional=True,
+        write_permission=View,
     ),
     TextField(
         name='marginalNotes',
@@ -2290,10 +2326,10 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            view to use it in the widget in edit mode.  This way, we can display
            more informations than just the title.'''
         if withTypeName:
-            return "{0} - {1}".format(translate(self.portal_type,
-                                                domain="plone",
-                                                context=self.REQUEST).encode('utf-8'),
-                                      self.Title(withMeetingDate=True))
+            portal_types = api.portal.get_tool('portal_types')
+            return "{0} - {1}".format(
+                portal_types[self.portal_type].title,
+                self.Title(withMeetingDate=True))
         return self.Title(withMeetingDate=True)
 
     def Title(self, withMeetingDate=False, withItemNumber=False, withItemReference=False, **kwargs):
@@ -2686,6 +2722,18 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         tool = api.portal.get_tool('portal_plonemeeting')
         cfg = tool.getMeetingConfig(self)
         return cfg.getBudgetDefault()
+
+    security.declarePublic('showField')
+
+    def show_field(self, field_name):
+        '''See doc in interfaces.py.'''
+        item = self.getSelf()
+        if not item.isDefinedInTool() and \
+            item.attribute_is_used(field_name):
+            # evaluate TAL expression
+            tool = api.portal.get_tool('portal_plonemeeting')
+            cfg = tool.getMeetingConfig(item)
+            return cfg.eval_tal_expr_for_field(item, field_name)
 
     security.declarePublic('showObservations')
 
@@ -4995,7 +5043,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     def addRecurringItemToMeeting(self, meeting):
         '''See doc in interfaces.py.'''
         item = self.getSelf()
-        wfTool = api.portal.get_tool('portal_workflow')
+        wf_tool = api.portal.get_tool('portal_workflow')
         tool = api.portal.get_tool('portal_plonemeeting')
         try:
             item.REQUEST.set('PUBLISHED', meeting)
@@ -5010,11 +5058,11 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 trs = cfg.getTransitionsForPresentingAnItem(
                     org_uid=item.getProposingGroup())
                 if "validate" in get_transitions(item):
-                    wfTool.doActionFor(item, "validate")
+                    wf_tool.doActionFor(item, "validate")
                     trs = ["present"]
                 for tr in trs:
                     if tr in get_transitions(item):
-                        wfTool.doActionFor(item, tr)
+                        wf_tool.doActionFor(item, tr)
             # the item must be at least presented to a meeting, either we raise
             if not item.hasMeeting():
                 raise WorkflowException
@@ -5029,8 +5077,18 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
     def _bypass_meeting_closed_check_for(self, fieldName):
         """See docstring in interfaces.py"""
-        if fieldName in ['internalNotes', 'marginalNotes']:
+        if fieldName in [
+                'internalNotes', 'marginalNotes',
+                'neededFollowUp', 'providedFollowUp']:
             return True
+
+    def _bypass_write_perm_check_for(self, fieldName):
+        """See docstring in interfaces.py"""
+        if fieldName in ['neededFollowUp', 'providedFollowUp']:
+            item = self.getSelf()
+            tool = api.portal.get_tool('portal_plonemeeting')
+            cfg = tool.getMeetingConfig(item)
+            return cfg.eval_tal_expr_for_field(item, fieldName, mode='edit')
 
     def _bypass_quick_edit_notify_modified_for(self, fieldName):
         """See docstring in interfaces.py"""
@@ -5057,10 +5115,18 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         # some fields are still editable even when meeting closed
         bypassMeetingClosedCheck = bypassMeetingClosedCheck or \
             self.adapted()._bypass_meeting_closed_check_for(fieldName)
+        bypassWritePermissionCheck = bypassWritePermissionCheck or \
+            self.adapted()._bypass_write_perm_check_for(fieldName)
+        # write_permission is "View" for custom management
+        # if bypassWritePermissionCheck is False, make sure write_permission
+        # is no more "View", set it to "Modify portal content"
+        write_perm = field.write_permission
+        if not bypassWritePermissionCheck and write_perm == "View":
+            write_perm = ModifyPortalContent
         res = checkMayQuickEdit(
             self,
             bypassWritePermissionCheck=bypassWritePermissionCheck,
-            permission=field.write_permission,
+            permission=write_perm,
             expression=self.Schema()[fieldName].widget.condition,
             onlyForManagers=onlyForManagers,
             bypassMeetingClosedCheck=bypassMeetingClosedCheck)
@@ -8070,7 +8136,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             dest_field = newItem.getField(dest_field_name)
             # check that we will not empty a required field (case for "title" especially)
             # and also that if field optional, it is used in destination config
-            if (self.fieldIsEmpty(other_mc_field_name) and
+            if (fieldIsEmpty(other_mc_field_name, self) and
                 self.getField(dest_field_name).required) or \
                (getattr(dest_field, 'optional', False) and
                     not newItem.attribute_is_used(dest_field_name)):
@@ -8558,6 +8624,46 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             return item_state in cfg.getPositiveDecidedStates()
         else:
             return item_state in cfg.getItemDecidedStates()
+
+    def may_view_follow_up(self,
+                           field_name='neededFollowUp',
+                           label_ids=('needed-follow-up', 'provided-follow-up'),
+                           suffixes=[]):
+        """Helper methods for default view access to followUp related fields."""
+        tool = api.portal.get_tool('portal_plonemeeting')
+        cfg = tool.getMeetingConfig(self)
+        if tool.isManager(realManagers=True):
+            return True
+        is_manager = tool.isManager(cfg)
+        # same condition for any field
+        # must have relevant labels and MeetingManager or proposing group member
+        if (not fieldIsEmpty(field_name, self) or
+            get_labels(self, label_ids=label_ids)) and \
+           (is_manager or tool.user_is_in_org(
+                org_uid=self.getProposingGroup(), suffixes=suffixes)):
+            return True
+
+    def may_edit_follow_up(self,
+                           field_name='neededFollowUp',
+                           label_ids=('needed-follow-up', ),
+                           suffixes=[]):
+        """Helper methods for default edit access to followUp related fields."""
+        tool = api.portal.get_tool('portal_plonemeeting')
+        cfg = tool.getMeetingConfig(self)
+        if tool.isManager(realManagers=True):
+            return True
+        is_manager = tool.isManager(cfg)
+        if field_name == 'neededFollowUp':
+            # must have relevant labels, only editable by MeetingManagers
+            if get_labels(self, label_ids=label_ids) and is_manager:
+                return True
+        elif field_name == 'providedFollowUp':
+            # must have relevant labels and be MeetingManager
+            # or proposing group editor
+            if get_labels(self, label_ids=label_ids) and \
+               (is_manager or is_proposing_group_editor(
+                    self.getProposingGroup(), cfg, suffixes=suffixes)):
+                return True
 
 
 registerType(MeetingItem, PROJECTNAME)