+++ src/Products/PloneMeeting/model/adaptations.py
@@ -8,8 +8,10 @@ from imio.helpers.content import object_values
 from imio.pyutils.utils import merge_dicts
 from imio.pyutils.utils import replace_in_list
 from plone import api
+from Products.CMFCore.permissions import AccessContentsInformation
 from Products.CMFCore.permissions import DeleteObjects
 from Products.CMFCore.permissions import ModifyPortalContent
+from Products.CMFCore.permissions import View
 from Products.CMFPlone.utils import safe_unicode
 from Products.PloneMeeting import logger
 from Products.PloneMeeting.config import AddAnnex
@@ -97,7 +99,6 @@ WAITING_ADVICES_FROM_STATES = {
 # defined here to be importable
 WAITING_ADVICES_FROM_TRANSITION_ID_PATTERN = 'wait_advices_from_{0}'
 WAITING_ADVICES_FROM_TO_TRANSITION_ID_PATTERN = 'wait_advices_from_{0}__to__{1}'
-WAITING_ADVICES_NEW_STATE_ID_PATTERN = '{0}_waiting_advices'
 
 # restrict item validation back shortcuts
 # if not empty, we will permit back shortcuts from given item states
@@ -899,8 +900,13 @@ def _performWorkflowAdaptations(meetingConfig, logger=logger):
             from Products.PloneMeeting.MeetingItem import MeetingItem
             edit_permissions = [ModifyPortalContent, DeleteObjects]
             for field in MeetingItem.schema.fields():
-                if field.write_permission and field.write_permission not in edit_permissions:
+                # in some case we protect edit with "View" permission because
+                # we manage access manually, ignore these edit permissions
+                if field.write_permission and \
+                   field.write_permission not in edit_permissions and \
+                   field.write_permission not in (View, AccessContentsInformation):
                     edit_permissions.append(field.write_permission)
+            new_state_id_pattern = '{0}_waiting_advices'
             # try to get meetingConfig id in WAITING_ADVICES_FROM_STATES
             # if not found, look for a "*" that is applied to every meetingConfigs
             # else nothing is done
@@ -935,9 +941,9 @@ def _performWorkflowAdaptations(meetingConfig, logger=logger):
                 # generate state any_validation_state_waiting_advices when using
                 # MeetingConfig.itemWFValidationLevels, else a custom __or__ name
                 new_state_id = infos.get('new_state_id', None) or \
-                    (WAITING_ADVICES_NEW_STATE_ID_PATTERN.format('any_validation_state')
+                    (new_state_id_pattern.format('any_validation_state')
                      if item_validation_states
-                     else WAITING_ADVICES_NEW_STATE_ID_PATTERN.format('__or__'.join(from_state_ids)))
+                     else new_state_id_pattern.format('__or__'.join(from_state_ids)))
                 if not new_state_id.endswith('_waiting_advices'):
                     raise Exception('Waiting advices "new_state_id" must end with "_waiting_advices" !')
                 back_transition_ids = []