+++ src/Products/PloneMeeting/utils.py
@@ -2345,9 +2345,9 @@ def get_last_validation_state(item, cfg, before_last=False, return_level=False):
         return res
 
 
-def is_proposing_group_editor(org_uid, cfg):
+def is_proposing_group_editor(org_uid, cfg, suffixes=[]):
     """ """
-    suffixes = cfg.getItemWFValidationLevels(data='suffix', only_enabled=True)
+    suffixes = suffixes or cfg.getItemWFValidationLevels(data='suffix', only_enabled=True)
     return cfg.aq_parent.user_is_in_org(org_uid=org_uid, suffixes=suffixes)